Refactor flashcard display in HighlightViewModal
- Rename FlashcardListCard.tsx to FlashcardCard.tsx for better naming consistency
- Add showSourceHighlight prop to FlashcardCard to optionally hide source highlight preview
- Remove <li> wrapper from FlashcardCard component (moved to parent)
- Update FlashcardSection to use FlashcardCard component instead of inline implementation
- Implement two-column grid layout for flashcards in HighlightViewModal
- Add edit functionality to FlashcardSection with update mutation support
- Pass full highlight object to FlashcardSection for complete context
